conversion

+++ cpp/tests/test_ide_seir.cpp
@@ -27,15 +27,15 @@
 
 TEST(TestIdeSeirMin, simulateDefault)
 {
-    double tmax = 1;
-    double dt   = 0.1;
+    int tmax  = 1;
+    double dt = 0.1;
 
     using Vec = mio::TimeSeries<double>::Vector;
 
     mio::TimeSeries<double> init(1);
-    init.add_time_point<Eigen::VectorXd>(-12.0, Vec::Constant(1, 10));
+    init.add_time_point<Eigen::VectorXd>(-12.0, Vec::Constant(1, 10.));
     while (init.get_last_time() < 0) {
-        init.add_time_point(init.get_last_time() + dt, Vec::Constant(1, 10));
+        init.add_time_point(init.get_last_time() + dt, Vec::Constant(1, 10.));
     }
 
     mio::iseir::IdeSeirModel model(std::move(init), dt, 10);
@@ -44,7 +44,7 @@ TEST(TestIdeSeirMin, simulateDefault)
 
     // TODO: It is actually an inconsistency that the simulation of the IDE model goes to tmax + dt,
     // instead of tmax; this has to be corrected in the future.
-    EXPECT_NEAR(result.get_last_time(), tmax + 0.1, 1e-10);
+    EXPECT_NEAR(result.get_last_time(), (double)tmax + 0.1, 1e-10);
 }
 
 class TestIdeSeir : public testing::Test
